requirements.txt
fastapi
uvicorn
sqlalchemyz
pydantic
passlib
python-jose
python-multipart
streamlit
requests
pandas

